Separate rpc enums into specific categories
In order for a state machine to check that it has received a valid
message, each enum variant should have a unique determinant value. As
well, categories of request/error/notification rpc enums help to
maintain allowable and acceptable messages.

This patch separates rpc enums into 3 categories:

1. General (non-versioned)
2. Session (versioned)
3. Protocol (versioned)

The General category is unversioned and is only tied to the overall sasd
semantic version. Session and Protocol categories are tied to both an API
level and the sasd semantic version.