New module: protocol
Adds a new module named protocol. This module implements a new State
trait. This trait has the property where it will always be able to
handle a Version request or a Done notification. If a message is not one
of these, then the trait dispatches to the concrete state
implementation.

Also included is a new struct called Start which implements the new
State trait. This struct cannot handle any messages at all, and will
error out if a message is not handled by the State trait.