Enable platform specific states
The security mechanism on windows is different than on unix-like systems
due to the absence of Unix Domain Sockets on windows. As named pipes are
a security hazard due to the various ways an attacker can gain access to
a named pipe, the windows platform must implement an additional state in
order to verify incoming connections.

This change enables using different states for incoming connections
based on what platform the server is on.