Implement loading configuration data
This patch implements the loading of settings from a configuration file.
While the config file format is not defined in this patch, it's expected
that it will be in TOML format.

Please note that only a single integration test exists in this patch.
This is being commited as another feature depends on this. Full tests
will be created in a future patch.