windows: new alternate state InitSession
On unix-like platforms, the ability to use Unix Domain Sockets is enough
to verify that a client is being run by the same user as that running
the server (ie can restrict the socket on the filesystem to be readable
and writable by only a specific user). On Windows, however, there is no
such thing as Unix Domain Sockets, so a minimal authentication protocol
needs to be run before allowing access to the server's process-level
session.

As the winapi crate currently is missing some key api function calls to
be able to facilitate user identification, the InitSession state instead
resorts to creating a hidden file in a specified location on the
filesystem and asks the client to reply back with the token that's
contained inside this file. Once the state machine is built, it is
intended that this file remain open only for a short period of time
(10s? 30s?) after which it is automatically deleted. Also, while the
file is open, it can only be read from, no other process may write or
delete it.