windows: new state AuthSession
On windows, after a client has requested to connect to a session, the
client needs to provide an authentication token to prove that the client
is running as the same user as the user running the server. This is done
by the server creating a temporary file that holds the auth token and
requests that the client read this file and reply with the contained
auth token.

The AuthSession state is responsible for comparing the auth token that
the client provides with the expected auth token.

Unfortunately, this security measure is quite fragile, especially since
there is no easy way to manipulate file permissions on windows. Once the
winapi crate has reached version 0.3 (which is the next version
that joins all previously separate winapi crates into a single crate),
and the full breadth of the windows api becomes available, a different
procedure will be implemented. Until then, it is not recommended that
this program be used in production on windows.