windows: authenticate using only a single auth token
This change removes generating a client token by the server, and also
removes requiring a client to provide the client token when
attempting to connect to the server session.

The client token is largely useless; it was originally meant to
distinguish between multiple client sessions per tcp connection, but
as sasd development progresses, now there is only going to be a single
client session per tcp connection. As such, there is no longer any need
for the client token, and the auth token suffices to determine if the
client is being run by the same user as the server.