Remove use of trait objects from state transitions
While the use of trait objects made it easy to store a state, it
requires an allocation onto the heap for every state transition. For
example, if state A transitioned to state B, B would be boxed in order
to store B as the current state in the state machine.

While this might be acceptable when running the program a single time,
it may not be acceptable in high load situations. In addition, the use
of trait objects means that no optimizations can be made on any state
transition.

To remove trait objects, the StateKind enum is repurposed into a
StateValue enum, and each variant of StateValue represents a specific
state. Thus every state transition will then return a specific
StateValue variant. This makes it easy for optimizations to occur while
retaining the ability to store any state in the state machine.